+++ lib/replay.c
@@ -560,7 +560,7 @@ static bool rm_parrot_check_types(RmCfg *cfg, RmFile *file) {
     case RM_LINT_TYPE_BADUGID:
         return cfg->find_badids;
     case RM_LINT_TYPE_UNIQUE_FILE:
-        return cfg->write_unfinished;
+        return cfg->hash_uniques;
     case RM_LINT_TYPE_PART_OF_DIRECTORY:
         return true;
     case RM_LINT_TYPE_UNKNOWN: